+++ src/Config.h
@@ -34,8 +34,8 @@
 #define VER_LAMEXP_MINOR_HI					1
 #define VER_LAMEXP_MINOR_LO					4
 #define VER_LAMEXP_TYPE						Alpha
-#define VER_LAMEXP_PATCH					1
-#define VER_LAMEXP_BUILD					1856
+#define VER_LAMEXP_PATCH					4
+#define VER_LAMEXP_BUILD					1866
 #define VER_LAMEXP_CONFG					1818
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -50,10 +50,9 @@
 ///////////////////////////////////////////////////////////////////////////////
 
 #define VER_LAMEXP_TOOL_NEROAAC				1540
-#define VER_LAMEXP_TOOL_FHGAACENC			20120806
+#define VER_LAMEXP_TOOL_FHGAACENC			20151024
 #define VER_LAMEXP_TOOL_FDKAACENC			62
-#define VER_LAMEXP_TOOL_QAAC				244
-#define VER_LAMEXP_TOOL_COREAUDIO			7980
+#define VER_LAMEXP_TOOL_QAAC				258
 
 ///////////////////////////////////////////////////////////////////////////////
 // Helper macros (aka: having fun with the C pre-processor)