+++ src/Encoder_FLAC.cpp
@@ -109,6 +109,11 @@ public:
 		static const char* s_extension = "flac";
 		return s_extension;
 	}
+
+	virtual bool isResamplingSupported(void) const
+	{
+		return false;
+	}
 }
 static const g_flacEncoderInfo;