+++ src/Encoder_Vorbis.h
@@ -37,7 +37,6 @@ public:
 	virtual bool encode(const QString &sourceFile, const AudioFileModel_MetaInfo &metaInfo, const unsigned int duration, const QString &outputFile, volatile bool *abortFlag);
 	virtual bool isFormatSupported(const QString &containerType, const QString &containerProfile, const QString &formatType, const QString &formatProfile, const QString &formatVersion);
 	virtual void setBitrateLimits(int minimumBitrate, int maximumBitrate);
-	virtual void setSamplingRate(int value);
 
 	//Encoder info
 	virtual const AbstractEncoderInfo *toEncoderInfo(void) const { return getEncoderInfo(); }
@@ -47,5 +46,4 @@ private:
 	const QString m_binary;
 	int m_configBitrateMaximum;
 	int m_configBitrateMinimum;
-	int m_configSamplingRate;
 };